Extend copyrule to handle input-output attributes
In addition to computing and adding missing input attributes to
nonterminals, we also check to see whether the missing attribute is
also an output attribute.  If so, we make sure to add it to the output
attributes of the nonterminal if necessary.

This lets us thread input-output attributes through rules, for
example, in this grammar

   A@(;x:int)>@(;x:int) = ...
   B = ... A ... A ...
   C = ... B@(;x=3) ...

the attribute x is an input-output attribute.  It does not need to be
explicitly mentioned in B, the copyrule transforms this to

   A@(;x:int)>@(;x:int) = ...
   B@(;x:int)>@(;x:int) = ... A@(;x=x) ... A@(;x=x) ...
   C = ... B@(;x=3) ...

The attribute x is not considered missing from C, so C's definition is
not affected.  However, x is missing from B, so it becomes an
input-output attribute of B.  Notice that the value of x for the first
occurrence of A in B gets its value from the input attribute of B, and
the second gets its value from the output x of the first occurrence of
A.