Added grammar for lexer-based ocaml grammar.
This grammar is a hand-modified version of one auto-converted with
our ocamlyacc converter tool. It is in progress. At this point, it
can successfully parse some ocaml files. I am still fiddling with
the precedence rules to try to increase filtering to the point of
matching the parse produced by the real ocaml grammar.