Added new precedence analysis based on Visser.
The new precedence analysis is a direct implementation of the ideas in Visser's thesis. It is far simpler than the earlier analysis. Basic testing indicates that, while it doesn't improve the performance on the OCaml grammar, it only degrades by less than 10%.  Clearly, more work is needed, but this is a start. This formulation is nearly identical to dypgen's. The main difference is that we do more inference -- specifically, the annotations for left and right subcomponents of productions with precedence.

I also updated the ocamlgrammar to use the new system.