Added SPPF printing.
I've added two functions to history to print an SPPF. One prints a single parse tree and reports ambiguities. The other prints the entire SPPF, complete with loops and shared nodes. Currently, only the latter is exported from the history module. These functions are still under construction, but are functional enough for a commit.