Fixed history memoization problems.
At a high level, the history is memoization is now keyed by a more specific key, thereby doing a better job of differentiating different parses. This change actually affects more than just memoization. It also solves the problem of right recursion loops even with memoization off.

In more detail, history now uses the starting position of a subparse as part of the node key. Also, I've revisited the choice of what exactly is memoized. Previously, the Root data structure was memoized. I've changed this so that instead the info data structure is memoized. I've also changed the history functions to take a position parameter. Ideally, only empty would need this parameter, but then every call in the grammar would need to be parameterized and invoke empty, thereby destroying call-collapsing. So, instead, i've hacked things so that the first push or merge will provide the starting position.