Upgraded History interface to hide more implementation details.
I changed the interface of the History module to hide some more of the internal details of the history and postfix classes. In the new interface, only class types are exported. The history class type is created via the (pre-existing) factory method 'new_history' and the postfix traversal is now created by a method call on a particular history.

Finally, I upgraded the ocamlparser.bnf file to use the new interfaces.