Exposed label abstraction in history type.
The relevant types in History are now parameterized by a label type. For the time being, we're stilling instantiating this type with a type from within the History module. So, this is just step one. The next step is to instantiate with a client-specified type.

In addition to the changes to History, I updated all modules using History in the obvious (uninteresting) way.