Added support for inspecting semantic values in the Earley graph.
The most significant change comes in wfe.ml. I've added a function to fold over all live
semantic values in the Earley graph. I've included a simple application of this fold to
count the number of live semantic values reachable from the current Earley set. I've also
added a new field to the SEMVAL signature so that client-specified folds can
be performed.

I've updated code-generation in main.ml to include default values for the folding functions
for all grammars, and updated the generated code accordingly. These additions are only step 1.
Next, I plan to add flags to allow the generation of interesting folding functions for grammars.