Added inspection of histories
I changed the History module to export an interface for collecting sets of root 'id's, where the ids are opaque. I added code to the codegen in main.ml to use this feature and the engine's inspection feature together to count the number of unique roots in play at any given time during parsing.