Late positions now supported by dispatch
We no longer desugar $pos and we take advantage of
this in the Lift phase when binding a substring, e.g.,

    "abc"$x

This reduces the need for coroutines.  One common
reason we still need coroutines is that we do not
do unrolling of *, instead we use a counter.
Unrolling seems broken right now.

Fixed printing of relevance.

Workaround for poor parsing of .bnf files:
we do not properly parse accumulating star
that contains brackets, so I added Yak.Util.nil.