Bug fix for dot_show for histories
The dot_show function for histories assumed that the history forests (the roots) were built with memoization. Specifically, in its internal table used to make sure that the same node was not processed twice, it compared nodes via their labels. This comparison is wrong on two counts. First, if the nodes have not been built w/ memoization, then a shallow comparison of this type is wrong, and either a deep comparison or physical equality should be used. If they *have* been built with memoization, then physical equality is sufficient.

To fix, we change to building the hashtable using physical equality. This corresponds to marking the actual node when its visited, so works out nicely.