Added unambiguous versions of box syntax.
The standard version is ambiguous -- the inside-text of the box code conflicts with the optional return type and nullability declaration. By luck (or not), the previous version of the Yakker parser (w/o memoization) disambiguated correctly. With memoization turned on, the boxes with
return types are parsed incorrectly.

I have added two more versions of boxes, for experimenting. Once we settle on a preferred syntax, we can replace the old box and update all tests/examples appropriately.