Added flag and support for inlining nullable nonterminals.
I have added front-end support for inlining the null grammar of nullable nonterminals.
This design provides an alternative to dynamically checking for nullability
immediately following calls. It should drammatically simplify one of the trickiest
parts of the Earley engine.

As part of this change, i've added a flag and command line control of the flag.
Nullable nonterminals are inlined now by default. Long-term, however, I hope
to remove any alternative and simplify the engine correspondingly.

Multiple modifications in Fsm module.
1) Added support for new Gil When_special construct. Now translated into EPredSpecial,
   which, in turn, generates a WhenSpecial instruction.
2) Added support for inlining nullables, controlled by new flag.
3) Moved *already_var* function into Variables module.
4) Various simple code/comment cleanups/mods.
5) Moved generation of nullability predicates above generation
   of variable bindings used in transducer instructions. This move
   was prompted by the fact that inlined When_special code can refer
   to nullability predicates, hence they have to appear earlier.