Updated nullable-nonterminal inlining to handle lookahead case.
Previously, the right-side included in a Lookahead element was left unchanged during inlining. This update fixes that oversight and modifies the lookahead properly. Note that this new version can result in lookahead right-sides which do not meet the lower-level constraints on lookahead. Hence, it should be followed by a desugaring.