Added new version of Fusion which uses HOAS to build fused expressions.
The new version has been added to the same file as the original, and the original has been left intact. We can remove during the next cleanup sweep.  The new version, in addition to building expressions in HOAS, performs a simplification on built expressions before returning. So, in principle, it might produce reduced expressions w.r.t. to the earlier version.