New det. branch construct and misc. cleanup.
This commit introduces a new language construct: a deterministic branch. It allows for matching an expression against a datatype constructor. Moreover, it is designed so that multiple (disparate) branches on the same type can be merged into a single pattern-match in the backend.

In order to get this benefit, though, we need to ensure that branches on the same value can be factored/determinized together. Towards this end, I've also added a new labeling transform that tries to assign identical labels to identical right-sides to improve the impact of determinization.

For more details about this construct, see the blog. Particularly,

    https://yakker.research.att.com/blog/yitzhak/Proposal-for-deterministic-branching-construct-201102021553.txt

I have modified all phases to handle the new construct. However, for a few cases which were not immediately relevant, I simply added a todo, which we should return to in the future.

Since the new construct includes a type annotation which is relevant for wrapping, I've changed the wrapping transform to traverse right sides to collect types for the table. Previously, only the parameter and return-types of rules were included in the table.

I have not added external syntax for this new construct. Instead, it can only be introduced by a (new) compilation from declared lexical tokens.

During the course of all these changes, I also made a few miscellaneous (and small) code cleanups, which are also part of this commit.