Added simpler version of deterministic branching.
I've added a new version of deterministic branching that does not involve early or late actions. It bundles a full box with the branch construct which is executed and then pattern-matched on, so no binding is needed. Also, the results are discarded -- they are not available for either early or late binding. This new version is intended largely for experimentation. A practical construct will require the ability to save the data carried by some tokens.

The new version overloads the Gul and Gil DBranch construct, but uses a new PAM instruction and engine instruction.

To use the new version, set Compileopt.late_only_branch to true.

In addition, I've refined the code for generating nullability predicates somewhat to reduce the number of dead predicates generated (i.e. those never used).