(Partially) generalized engine w.r.t. to terminal processing.
I have further modified the handling of tokens in the engine to treat them (almost) on par with actual characters. In this new version, use of a lexer is incompatible with scannerless parsing -- tokens always advance by one Earley set, and the position is managed by the lexer. As a result, tokens are now considered 'never-nullable'. Furthermore, the _parse function is now parameterized by assorted terminal-stream management functions: advance, fill, save, restore.

The savings from this change are negligable. This observation implies that the cost of the generality of boxes + scannerless might be less than I anticipated.