Added specialized token lookahead.
I have added a special case for lookahead of a single token to avoid performing a full-blown lookahead (which involves a recursive call to the _parse function, which is quite expensive). This change had a very significant impact for the ocaml parser. It brought down the cost of parsing the stdlib from ~4.6 seconds to 0.23 seconds. This new speed is about twice as fast as scannerless parsing for the same grammar.