Added tokenizer syntax and cleaned up codegen.
The most significant change is that there is now syntax to specify the use of a single lexer/tokenizer, in place of scannerless and/or multi-lex based parsing. I have added the '@set-lexer ...' concrete syntax, and corresponding abstract syntax, to invoke this feature. In addition, as part of this change, i've improved the codegen so that it is fully automated.