(Subtly) changed the semantics of lexstate used in engine for single lexer mode.
In particular, changed the meaning of the boolean in the lexstate tuple to make the code clearer. Changed from indicating whther the buffer is at EOF to whether it is *not* at EOF (scanning is possible). This change affects both the engine and the support code in YkBuf.

In addition, for improved error reporting in lexer case, disabled code which backed up the buffer by one character after an error. Unfortunately, this change can break error reporting for scannerless grammars, so we need to implement a better solution which handles both correctly.