Abstracted step_back function in engine.
Scannerless parsing and tokenized parsing need to step back in different ways (used in the case of an error). Formerly, we hardcoded a call to YkBuf.step_back, which worked for scannerless but broke tokenized. Then, I disabled that call, which reversed the brokenness. I've now added it to the TERM_LANG interface, so that each term language can handle it appropriately. The default scannerless TERM_LANG implementation just calles YkBuf.step_back. The default tokenized functor just puts in a dummy function -- we need to add proper support down the line.