Move reversal code from Main to Replay
It is better to put the reversal code into Replay because sometimes a
grammar is late-relevant but does not require a history, e.g., a
grammar with late actions but no Alt or Star.  In such cases no
reversal is required either.  By generating the reversal code in
Replay only as necessary, we avoid having to expose this to Main.