Add a command to output the result of the replay transformation
The replay transformation is the last before dispatch.

The output of the replay transformation includes @delays corresponding
to history outputs.

All of the inserted delays involve closed expressions, and are
currently printed as @-delay(...).  At the moment this syntax is not
recognized by the yakker parser, so we cannot use this as the input to
yakker (e.g., with -after).