Optionally use prefix histories for replay
The current default history mechanism outputs labels as late as
possible, e.g., transform

    ("a" {...} | "b" {...})

to

    ("a" :l1 | "b" :l2)

and

    *( r )

to

    :l_done *( r :l_body)

Hopefully this leads to improved left factoring (more call
collapsing).  We call this the "postfix" history stategy.
It requires "reversing" the history for replay.

Previously we used a "prefix" history strategy, e.g., transform

    ("a" {...} | "b" {...})

to

    (:l1 "a" | :l2 "b")

and

    *( r )

to

    *(:l_body r) :l_done

This decreases left factoring opportunities but does not require
"reversing" on replay.  A tradeoff, so we support both, postfix is the
default, use the -prefix-history to get prefix.