Fixed a replay bug and a tyspec bug
I fixed a bug in replay wherein the mechanism used to save positions resulted in unintended node sharing in the resulting SPPF. The problem relates to fact that every node in the SPPF needs to be uniquely labeled w.r.t. to the grammar. That is, nodes corresponding to two different points in the grammar must have different labels. This policy guarantees that an sppf node is only shared if it corresponds to the same point in the grammar for the same input span.

The previous version of replay violated this requirement by using the position itself as part of the node label. As a result, different pushes relating to the same input span but *different grammar positions* shared the same SPPF node.

To fix this bug, I have modified the way positions are saved and recovered. Each position-push in the grammar now has its own label, just like a standard push. Indeed, Position false is now rewritten to be a normal push via Delay. Then, in the replay functions, when a position is needed the function _p is called rather than _n. Correspondingly, a _p function is defined and added to all replay-function calls/declarations.

Note that the same problem encountered here for positions applies to the saving of any arbitrary data via Delay. For now, we ignore this issue, special-casing position. We leave general Delay for future work.

replay.ml: the changes discussed above

Makefile: removed test_tyspec, the ad-hoc test for this bug, and added a new regression oldtyspec which captures the test. Note that a bzr up is needed to get the new regression.

tyspec.bnf: Previously, I was only mapping type variables which could not be generalized. However, *all* type variables need to be mapped, as we need a fully concrete type for sv. I have updated the code in tyspec to drop the disctinction between type variables, and just map all of them. That is, anything starting with a single quote will be mapped.

tyspec.ml: Incorporates the above changes and the modifications to replay codegen.