Added support for Delay to arrow-notation.
I've added support for coexisting with late actions to the arrow notation transformation. Much like dispatch, it checks for the various different combinations of early/late. Ultimately, this support is temporary until we change replay to be handled entirely as a Gul-to-Gul transformation.

The bulk of the changes are in dearrow.ml.  These changes also include various bug fixes based on bugs which turned up in the testing process but not necessarily related to the feature being implemented. I also cleaned up the code (and process) dealing with inferring the type parameters for the ev type, including shifting code from Main to Dearrow. More generally, these changes appear in main.ml, fsm.ml, nullable_pred.ml.

In addition, in Main, I eliminated all but top-level use of the mutable variable outch,
and, in the dispatch phase, I moved producers and relevance analysis so as to also
be executed for arrow-notation.