Modified history functionality to properly support delay of arbitrary values.
Previously, we were improperly/insufficiently labeling forest nodes that contained arbitary values. I have significantly reworked the replay mechanism so as to guarantee that all nodes are properly labeled. The changes are both at the meta level -- the codegen pipeline itself is subtly, yet significantly, different -- and at the object level -- I have changed the shape of history data and the functions that manipulate it.

I have also added support for arbitrary delay to Dearrow. Previously, it was only available, if a bit brokenly, to coroutine transform. The new Dearrow can now handle all regressions (except t007), examples, tests and yakker_grammar itself.

The new version of support for arbitrary delay is a first-cut solution at this point. While I have high confidence that it works, I believe there to likely be significant room for optimization.

The highlites of the new mechanism as compared with the previous one are as follows:

   * Every datum pushed/merged into the history is
     labeled. Correspondingly, a strict distinction is made between
     labels and data. History elements now carry a triple of label,
     value and position.

   * Replay labels and history labels are now completely
     independent. Replay acts entirely at the level of Gul. Its
     Delay outputs (called labels) are opaque to the Gul implementation (Dispatch,
     Dearrow), and irrelevant. The only connection between the two is
     that they both use the name label.

   * Dispatch and Dearrow assume an invariant regarding Delay: all
     values inhabit the same value and are opaque. We no longer treat
     closed Delay as carrying a raw integer. Instead, it is wrapped
     (as needed) in Replay.

   * A nice side benefit of this approach is that the irrelevance of
     merge labels to replay becomes obvious -- they are entirely
     unrelated entities.

   * Normalized use of history constructors -- they are now the only
     API for performing history actions; no more direct calls on the
     history object.

   * For arrow notation case, I have added labeling to the pipeline
     for use by the new mechanism, which depends on labels.

In the process, I also fixed some bugs/made some improvements, including

   * removal of code in Replay's history-wrapping for late positions
     which had been obviated by the special-case handling of
     positions;

   * fixed handling of positions in history reversal when history
     values are wrapped; also cleaned up the design slightly via
     push_pos function to hide wrapping.

   * removed dead _d_and_push combinator from Early-Late prologue.

   * changed replay mk_... combinators to use labels from Label
     transform rather than rolling their own.

   * In Dearrow, generalized early-late mk_push to handle open values.

   * In Dearrow, fixed bug wherein Action(Some _, Some _) was not
     handling only the late action.