Refined context manipulations in Dearrow
In order to fix a number of bugs, I've reworked the handling of contexts specifically in regards to nonterminals. This issue arises in two settings -- when handling the Return_bind case of the bindq type and when transforming nonterminals.

Regarding the former, I've added a flag to Return_bind to indicate whether the nonterminal is a producer. I then use this flag to decide whether or not to return a value when Return_bind is encountered (as opposed to always returning a value, and using unit when no other appropriate value was available).

Regarding the latter, in generating merge code, we take into account the "producerness" of the nonterminal being processed, assuming that if it is not a producer then it will not return a value.

In addtion, I've changed the generation of the pattern for the empty environment. Specifically, I special-case handling of empty context to result in a wildcard, which is fine because the constructor for the empty environment carries nothing. This change allows call-collapsing to work in cases where caller has a non-empty context but callee has no arguments. Without this change, call-collapsing will result in a match failure, as the code will attempt to match a non-empty environment.

I've dropped the practice of skipping irrelevant rhs, because they might have
out-attribute producing symbols. Note that the code for coroutines doesn't need to worry about this because out-attributes have already been desugared at that point. We handle them directly. Moreover, coroutines generation doesn't need to check for producerness because wrapping desugars it into actions, which trigger relevance. In contrast, we need to consider both "producerness" and output attributes in determining of what and when to add to the environment.