Added regresssions for testing arrow-notation transformation.
I've added a new directory to the *git* repo: regress. In this directory are now a group of arrow-notation targetted tests which I used to test the code I've just committed (found quite a few bugs). They should work as well for coroutines, but don't quite yet because they are missing type annotations.

I've correspondingly modified the Makefile to account for the new regresssion. Note that the Makefile specializes them to always be compile with the -arrow-notation flag because of the issue mentioned above. We should remove that specialization once coroutines support type inference.