Fixed replay bug; added new analysis
One aspect of Replay was making a decision based on the relevance of rhs when it should have been based on the 'producerness' of the rhs. I've added a 'producerness' analysis (is_rhs_producer) to Analyze, and updated Replay to use it in place of relevance.

I've also reformated small bits of Analyze.produce.