Fixed two bugs in Dearrow
Both bugs related to handling of symbols. First, I was using producerness to decided whether to history-merge a symbol. But that's wrong, because it neglects side-effecting symbols, which are perfectly valid for history. That has now been fixed and variables appropriately renamed.

Second, I had forgotten to take late-relevance into account in one case of gen_ret in envUpdateMerge, which led to the generation of incomplete histories. That has now been corrected.