Fixed Rcount case of lifting to not produce empty actions
When a repeated element is unrolled in lift, but it has not relevance, the result will usually have some empty actions (Action(None,None)) because early and late relevance are handled orthogonally.

I have changed the case for Rcount to avoid this issue by special casing the case of a totally irrelevant body.