Added new yakker command; switched to arrows for Yakker grammar.
The changes in main.ml and cmdline.* relate to adding a new "parse" command which just parses the input and then exits. It is useful for checking the syntactic well-formedness of a grammar and for testing parsing speeds.

The changes in the Makefile and yakker_grammar.ml related to switching the boostrapping of Yakker's grammar to use the arrow notation for speed improvements.

The change to fsm.ml is to reinstate call collapsing for arrows+replay to help with Yakker's parsing performance.