Added yakker command to add dummy late actions to a grammar.
This new feature inserts late actions that attempt to simulate building a parse tree. The actions themselves are empty; what is significant is the choice of their location. I've attempted to place the actions realistically to simulate the impact of parse-tree building actions on the resulting automaton (particularly in terms of lost sharing).

The command is called "pads-lift" because its semantics is closer to the way that PADS automatically builds an AST than Yakker's standard lifting. In particular, a sequence of elements is followed by an action insertion even if the last element is not late relevant. This corresponds to building a tuple of the (relevant) elements of the sequence, rather than corresponding to returning the last element, as standard lifting does.