Fixed arrow-notation codegen to work with new engine.
There are actually two fixes here. First, I changed the internal invocation of ocamlc to include the -rectypes option which is necessary because of changes to the engien.

The second fix is actually a continuation of an earlier fix.  The arrow-notation codegen used to wrap the automaton in an internal module. But, this turned out to have (unexplained) impact on ocamlopt compile times, in some cases causing it not to terminate. I therefore turned off the wrapping. However, it is necessary in some cases (which, I now realize, is why I used it to begin with -- not just for hygiene).  The problem is that some of the generated code has free type variables which are not generalizable. Hence, they must be assigned a monomorphic type or not exported by the module.

I chose the second approach, and implemented it by wrapping those elements in an internal module. However, since that didn't work, I've now changed the codegen to take the first approach. That is the change in nullable_pred.ml.