Fix bug in late relevance analysis
Yitzhak found this one:

If a nonterminal has a late parameter but no late actions (no uses of
the parameter) then it was not being marked as late relevant.

NB: cmdline.ml and tyspec.ml are changing because of a problem with
openfst: it is not minimizing transducers identically on different
operating systems.  It produces equivalent but not identical outputs.