Added nullability attributes for nonterminal definitions.
I have added a new attribute for nonterminal definitions: nullability. This attribute can be used to influence the nullability-predicate analysis. It is intended for definitions with complex uses of box that cannot be adequately handled by the analysis.

Many of the changes related to getting the grammar to the nullability-predicates code. In the pre-commit versions, it only received the gil definitions, which lack the attributes.

gul.ml: addition of attribute.
  unrelated: additional commenting of token_decl type and removal of (dead) commented-out code.

main.ml: adapting from using Gil definitions to Gul grammar.

nullable_pred.ml: incorporation of attribute into inference of nullability predicates.

yakker_grammar.bnf: added the new nullability field to attribute initializition code.