Added scripted control of logging in engine.
I normalized all uses of logging in the engine to use an appropriate macro. Then, I changed the LOG macro to wrap calls in assert rather than checking Logging.activated. This way, I can turn off logging with an option to ocamlopt. I little bit hackier than I'd like, but the best for the time being.

Correspondingly, I changed the Makefile to check for a DO_LOG parameter and handle it just like DO_PROF -- logging is on and the whole compilation process takes place in "logbuild", rather than the standard "build" directory.