Added support for user-driven error-recovery.
I've added an error recovery mode and a special "error" nonterminal for grammars.  When an parsing error occurs, defined
as reaching an input position for which *no* transitions can be made (the next Earley set is _empty_), the engine will
now search for a transition on the special error nonterminal. If any are found, they will be processed to see if a transition
on the current input terminal is possible.

This feature involved changes to transducer code generator (fsm.ml), the JIT (pamJIT.ml[i]) and to the engine itself.

Also, in the course of modifying the engine, I tweaked some code to (perhaps) improve performance.