Added new engine variation with hashtable-based Earley sets.
I've added yet another version of the Earley set data structures. This one is somewhere between the flat sets and the sparse sets. It is uses a mutable hashtable, where the hashtable maps states to socvas. So, it is both mutable and groups items by state, but iteration is more expensive (requires allocating a closure) and clear is not constant time.