Significant changes to nullability predicate processing
I've made two significant changes to nullability predicates:
1) The first change is bug/performance fix. We were computing nullability predicates twice, first for inlining, second for printing. Clearly, this is a waste of time in the front end. Worse, since the inlining came first, the second derivation of n.p.s for printing purposes was in fact including the inlined nullability predicates.  To fix both problems, I added another field to the grammar record for nullability predicates. Once they are computed, they are added to that field.

However, that change posed a challenge in terms of module dependencies because the field is declared in Gul while the type of the field comes from Nullable_pred. Since the latter already depends on the former, the new field creates a circular (type) dependency. So, I moved the entire DB_levels module out of Nullable_pred and into Gil, which, while less than ideal, is hardly terrible considering DB_levels.exp are an abstraction of Gil syntax.

2) I added (conservative) ambiguity-checking to the nullability predicate generation. For now, I've placed it in the rewriting function, although it is probably best in its own function.  For the purposes of this check, I added a StarE expression to the DB_levels.exp type so that I can properly check Kleene-closures for ambiguity.