Generalized type of nonterminals in Gil rhs.
I've added a type parameter to Gil.rhs for the type of nonterminal names used in the Symb branch.  While I expect that it will usually be set to string, the generalization simplifies transformations that might want to manipulate more complex forms of names (for example, to maintain seperate namespaces for freshly introduced nonterminal names and original names).