Added a fresh implementation of nullable-nonterminal handling.
The new implementation is based closely on the description in the new Yakker paper "Practical Dependent Parsing". In most ways, it is semantically equivalent to the existing implementation, and the new version is in essence cleanup/refactoring. However, it also has substantial differences.

The implementation compiles at this point but is completely untested. I intend to more fully document this new version once it has been integrated into the Yakker pipeline and tested.