Begun code cleanup.
I've begun assorted cleanup of the Yakker code. I created a new module Oqv (Of Questionable Value) into which to place code that is not being used and should be reviewed for deletion. As a first step, I moved all of the code from the previous implementation of nullability predicates into Oqv. I am only keeping it at all because the new code performs slightly worse for a few tests and I'd like to investigated that further.