Improved handling of nullability attributes on nonterminals.
Previously, Nullable_pred handled nonterminals with nullability attribute other than Unknown by either ignoring them or marking them as fail (depending on which grammar was being derived). However, this approach misses some optimization opportunities relating to the encoding of lexer calls with box.

In the new version, Always_null means that the nonterminal is mapped to 'fail' in the substantive grammar, and Never_null means it is mapped to 'fail' in the epsilon grammar. Otherwise, the nonterminal is processed.