Refactored Replay module so as to cleanup prologue generation.
I modifed all replay functions that touched the grammar's prologue to instead return strings. These strings in turn are collected and then put together into a replay prologue which is then returned to the caller. It is then the caller's responsibility to actually add the prologue string to the grammar's prologue.

The change accomplishes two things, in order of importance:
1) Clarifies the relationship between elements of Replay and changes to the prologue. It is now much easier to look at the code and understand exactly what gets added to the prologue from Replay and what inputs influence those additions.
2) reduces the side effects of Replay.transform.