Added two flags to Yakker (internally).
I've added two new flags to Yakker to control elements of code generation:
1. Eta-expand transducer bindings. This feature ensures that all of the generated support-bindings for the transducer are eta expanded into functions. This expansion forces all of the bindings to be values, which is important because it ensures that type inference will not result in ungeneralizable type variables. Currently, the implementation is a bit heavy-weight because it does not check first whether the string is already a function, it just blindly expands.
2. Memoize epsilon-grammar parsers. This feature already exists -- the modifications allow it to be turned off. The new code for handling epsilon-grammars forces said grammars not be left-recursive. Therefore, there's no longer a need for memoization. However, in some cases it could result in speedup (just like memoization in other contexts). However, in testing, it did not show any improvement to performance, and, for some cases, resulted in worse performance. Therefore, we will soon turn it off by default.

Both of these features are building towards another, more significant, improvement: removing the explicit inference of the ev type from the Yakker code. Without it, eta-expanding and disabling memoization become necessary. The former, because of the error mentioned (which is avoided if the ev type is specified, for complicated reasons relating to type inference); the latter, because the memoization uses specialized versions of the Hashtbl module, via the functor interface, and therefore require the ev type.

For this commit, the settings of both flags have been let at their earlier defaults. These settings will likely change in future commits.