Fixed bug in generating nullability predicates when memoization is turned off.
Previous code resulted in nullability predicates erroneosly not being emmitted because of change to meaning of ntcalled. The assumption was that ntcalled only effects memoization, but that assumption was false. Value ntcalled also effects printing of nullability predicates.

The new version of the code targets the memoization-relevant portion of the code only.